Adopt "Starter Library" Name
To avoid confusion with the legacy ALAN Library v0.6.x, and the other
ALAN libraries, all library implementations in this project are renamed
to "Starter Library" (see #7).

Library source folders are renamed to `StartLib` (without locale suffix)
and all README files are updated accordingly.

The libraries' source files and ChangeLogs are yet unaltered, and their
adaptation is postponed to later commits.