StartLib EN v0.1.0
Edit contents of `lib_en/` to create the first Beta of the English
Starter Library, version 0.1.0 (see Discussion #14).

The main module `std.i` was renamed to `StartLib.i`.
Comments referencing the old Library and its changes were removed.
Apart from that, the codebase is the same as that of Library v0.6.2.

The new CHANGELOG only tracks changes introduced since Library v0.6.2,
as found on the archived copy at ALAN Goodie.