StartLib EN v0.1.2
Edit `examine.i` by adding to EVERY THING the new `xDesc` str attribute
which can be (optionally) used to provide quick static descriptions for
the 'examine' verb, which has been edited to print the `xDesc` of its
parameter when its a non-empty string, and fallback to the default
"There is nothing special about..." message otherwise.

Update "Cloak of Darkness" to make use of the new `xDesc` attribute in
those instances that allow it.