Update EClint Scripts
New version of `validate.sh` and `git-hook-install.sh` which will only
invoke EClint on files that are tracked by Git or staged, respectively.

The new `validate.sh` script is a bit slower to execute than its prev.
version, but it won't produce false positives for untracked files.

The Git hook validator script is generally faster since it only checks
files which are staged, focusing on the commit changes instead of the
whole repository.