Automated Tasks
Add to Rakefile a custom `transcripts_folder_to_task()` method that can
parse a folder and automatically add to a task the required storyfile
and transcripts as dependencies, and then create all their file tasks
dependencies. It accepts three parameters:

* target_task   -- The target task for storyfiles and transcripts.
* target_folder -- The folder that needs to be scanned.
* dependencies  -- Required ALAN sources to build the storyfiles.

This method is for handling folders containing sample adventures or
test files which depend on a library and come with solutions files to
generate various transcripts.

Currently the method can only handle folder with a single adventure
where all solutions should be run against it, but soon it will be
updated so that in case of multiple adventures it will associate to
each adventure solution files that share its same basename according
to the pattern `<advname>*.a3s`, and associate dependencies accordingly.