Fix Worn Items MESSAGEs
In the `mensajes.i` module, fix the runtime messages that list container
contents so that worn items are followed by a "(being worn)" message in
Spanish which correctly handles the noun gender and number.

Add to wearables test adventure a special 'LIST <ACTOR>' verb to test
the `LIST` statement on actors.