Rake: Allow Defining ALAN Include Paths
To make our `_assets/rake/alan.rb` module reusable with other ALAN repos
introduce a new global variable `$alan_include` to control the path for
the `-include` option passed to ALAN during compilation (if any).

This allows each project to define its own include paths in the main
`Rakefile`, so that the `alan.rb` module can be shared as is.

Since the ALAN i18n directories are structured in a consistent manner,
the include path for every library is always `../Foundation`, regardless
of the library locale.