New Spanish Documentation
Add in `alan_es/docs/` a draft of the user guide for Spanish Foundation,
currently written in English. The document supports dynamic examples via
external ALAN adventures and solution files which are re-compiled and
generated at build time, handled via the latest Rake functionality.

Rake takes care of handling all documentation prerequisites, so whenever
changes are detected in the AsciiDoc sources, Spanish library sources,
the external ALAN examples or their solution files, the documentation is
rebuilt automatically to ensure that any changes to the source code are
reflected in the documentation.