Foundation ES v0.3.1
Add to grammar module supports for special feminine singular nouns that
take the "EL" article, by checking at initialization time whether
instances with `artículo "el"` are set to `Is femenina`, in which case
they take masculine articles but their other grammar attributes are
treated as feminine.

Remove some linguistically incorrect alternative syntaxes with 'DE' in
the `hablar.i` and `poner.i` modules.

Update the Spanish documentation with a new section on how to implement
feminine instances with "EL" article.

Update `Rakefile` to treat ALAN sources and transcripts of documentation
examples as requirements for building the docs.

In `vampiro.alan` fix the article of an instance which was had correct
gender but wrong number ("la" instead of "las").