Tweak Ticker Mechanics
Add a button to start and stop the ticker, so we can test for run-time
messages without the clutter of the ticker reports.

Move the ticker counter to the new button and get rid of the `my_game`
location. Stopping the ticker resets the counter. The start value is now
zero, not one, because I've tweaked the ticker event to first increase
the counter, then show its value, which might be better if we want to
implement a turns board to be examined in the future (just in case).