Foundation ES v0.3.2
Fixed some verbs that relocated objects into containers:

- `dar.i` — verb `dar`.
- `lanzar.i` — verb `lanzar_en`.
- `poner.i` — verb `poner_en`.

Since the action is not guaranteed to succeed, e.g. due to the `locate`
statement failing because of a `taking` clause in the target container
which would raise a `CAN_NOT_CONTAIN` Run-Time MESSAGE without aborting
the verb execution, after the `locate` we check whether the object is
effectively inside the target container before reporting success and
tweaking the wearability attributes.