Foundation EN v0.2.3
Fixed some verbs that relocated objects into containers:

- "give.i"   —>  'give'
- "put.i"    —>  'put_in'
- "throw.i"  —>  'throw_in'

Since the action is not guaranteed to succeed, e.g. due to the LOCATE
statement failing because of a TAKING clause in the target container
which would raise a CAN_NOT_CONTAIN Run-Time MESSAGE without aborting
the verb execution, after the LOCATE we check whether the object is
effectively inside the target container before reporting success and
tweaking the wearability attributes.