Translate and fix several messages
- meta.alan
Translate messages in the 'meta.alan' source adventure and add some more synonyms for the 'presionar' verb. Also delete the accent mark in 'bidónes': the correct form is 'bidones'.

- mensajes.i
Fix an issue with the CONTAINMENT_LOOP2 Run-Time message, which wasn't handling the number of 'estar' verb properly.

Improve translation of 'NO_UNDO' and 'UNKNOWN_WORD' messages because the old ones had an unclear meaning in Spanish.

- lanzar.i
Fix the GNA in 'lanzar' VERB. Now it shows the proper message.