New Library Scheme Document
Replace `VERSION_SCHEME.md` with `VERSIONING.adoc`, an updated version
of the article explaining how the library version scheme works, now in
AsciiDoc format since it will ultimately be included in the distribution
package in HTML format — possibly as part of a larger document.

Update all references to the old markdown document.